Always visit home page before login
django_selenium_clean.selenium.login() was creating a session cookie
without visiting any page first. It subsequently checked whether that
cookie had a domain; if it didn't have a domain, it visited the home
page and updated the cookie. Apparently this was done (as opposed to
always visiting the home page first) to save a few requests and run the
tests faster.

This worked on Firefox up to 41, which by default showed a blank page
when Selenium started the browser (resulting in a cookie without
domain). But Firefox 42, upon starting, visits some page in mozilla.org,
and the cookie became attached to the mozilla.org domain (instead of
localhost:8081 or something).