Overhaul - no global variable
SeleniumWrapper didn't always properly terminate and close the browser.
It couldn't be done in __del__, because Python does not guarantee it is
called on program termination, and it couldn't be done with atexit,
because other modules (in that case the selenium module is crucial) may
have already been destroyed when the atexit handler is called.

Therefore the logic of the application of changed. Selenium is started
at the beginning of the SeleniumTestCase and is closed at the end. This
means that the browser is closed and re-opened between test cases. There
is no longer a global selenium variable; instead SeleniumTestCase has a
selenium attribute.