Add a force_login method to SeleniumWrapper
This commit adds a new force_login method to the SeleniumWrapper class.
This method is analogous to the Django test client method force_login
added in Django 1.9.

Unlike .login(), the .force_login() method skips authentication and
verification steps and directly sets the user to appear as logged in. In
particular, this allows inactive users to login.

The code for the login() and new force_login() method now both use a
private method _login(). This is analogous to the appropriate django
source code.

See also https://docs.djangoproject.com/en/1.9/releases/1.9/#tests